event knows whether it can accept registrations
Adds #accepting_registrations? onto Event, moving logic from the
controller into the model. Using rails date helpers to determine if the
registration end date is in the past or not.

The logic should remain the same, which means that you cannot register
on the registration end date, not sure if this is correct.

The test just uses plain rspec as I just don't get/like the
Given/When/Then stuff.